[fix] crash when no internet
This is caused by the IOException thrown by URLConnection.connect when
no internet is available.

Instead, enclose it within try-catch block

+++ app/src/main/java/studio/saladjam/jsonplaceholder/fragments/SelectImageFragment.kt
@@ -41,7 +41,7 @@ class SelectImageFragment : Fragment() {
                         binding.imageSelected.apply {
                             alpha = 0f
                             setImageBitmap(it)
-                            animate().alpha(1f).duration = 500L
+                            animate().alpha(1f).duration = 500.toLong()
                         }
                     } ?: binding.imageSelected.setImageResource(R.drawable.image_placeholder)
                 }
